Student Grades Manager
PROJECT:   Student Gradebook
    • Input student names and scores.
    • Calculate averages and determine letter grades.
    • Print class summary (highest, lowest, average).
SKILLS:    This shows data handling, conditionals, and iteration.
EXTRA:    I added median, range, and letter variations (A+, A, A-) 

NOTES:    In this project I wanted to display more complex usage of 'def' with the use of global, return keywords. Also the 'try' & 'except' blocks as well as avoidng the 'menu' system I have used often in prior projects which allowed me to solve my problems with a different approach.