Weather App using API
Weather App using API

this is my first experience with APIs in Python. I think this is a great tool and something I want to play around with more in the future. Endless possibilities. Currently the base code is looking good. I need to fix some issues with the 'weather_info' dictionary items being called in the final results (step 3). 

Completed in this commit:
   -   Gained an understanding of how to use the 'requests' module.
   -   Learned about HTTPS status codes
   -   Set up the base code for this project


NEXT STEPS:
   -   Fix the dictionary key/value issue that is currently present in the project


POSSIBLE ADD-ONS:
   -   Include an option to use users ZIP CODE as an option to get the weather